add soft_bodies types

+++ include/physics/spring.h
@@ -1,9 +1,10 @@
 #ifndef SPRING_H
 #define SPRING_H
 // spring connects points with id=A with id = B;
+#include <cstddef>
 struct Spring {
-    int idA;
-    int idB;
+    size_t idA;
+    size_t idB;
     float resting_length; 
     float stiffness;
     float damping;
